# forge-go/api/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LaunchPreflightRequest": {
        "description": "Request to evaluate the exact blueprint launch plan before launch.",
        "properties": {
          "configuration": {
            "additionalProperties": true,
            "default": {},
            "title": "Configuration",
            "type": "object"
          },
          "description": {
            "anyOf": [{"type": "string"}, {"type": "null"}],
            "title": "Description"
          },
          "guild_id": {"type": "string", "title": "Guild Id"},
          "guild_name": {"type": "string", "title": "Guild Name"},
          "org_id": {"type": "string", "title": "Org Id"},
          "user_id": {"type": "string", "title": "User Id"}
        },
        "required": ["guild_id", "guild_name", "user_id", "org_id"],
        "type": "object"
      },
...
      "LaunchRequirementAction": {
        "properties": {
          "href": {"type": "string"},
          "kind": {"type": "string"},
          "method": {"type": "string"},
          "requires_client_credentials": {"type": "boolean"}
        },
        "required": ["kind", "method", "href"],
        "type": "object"
      },
      "LaunchRequirementSource": {
        "properties": {
          "agent_id": {"type": "string"},
          "agent_name": {"type": "string"},
          "origin": {"enum": ["agent", "profile"], "type": "string"},
          "profile_key": {"type": "string"},
          "profile_name": {"type": "string"}
        },
        "required": ["agent_id", "agent_name", "origin"],
        "type": "object"
      },
      "LaunchRequirement": {
        "properties": {
          "action": {
            "anyOf": [
              {"$ref": "#/components/schemas/LaunchRequirementAction"},
              {"type": "null"}
            ]
          },
          "id": {"type": "string"},
          "kind": {"enum": ["secret", "oauth"], "type": "string"},
          "label": {"type": "string"},
          "missing_scopes": {"items": {"type": "string"}, "type": "array"},
          "optional": {"type": "boolean"},
          "scopes": {"items": {"type": "string"}, "type": "array"},
          "sources": {
            "items": {"$ref": "#/components/schemas/LaunchRequirementSource"},
            "type": "array"
          },
          "status": {"enum": ["configured", "missing", "insufficient_scope", "provider_unavailable"], "type": "string"}
        },
        "required": ["id", "kind", "label", "optional", "status", "sources"],
        "type": "object"
      },
...
      "LaunchPreflightResponse": {
        "properties": {
          "expires_at": {"format": "date-time", "type": "string"},
          "fingerprint": {"type": "string"},
          "id": {"type": "string"},
          "ready": {"type": "boolean"},
          "requirements": {
            "items": {"$ref": "#/components/schemas/LaunchRequirement"},
            "type": "array"
          },
          "status": {"enum": ["ready", "blocked"], "type": "string"}
        },
        "required": ["id", "fingerprint", "status", "ready", "expires_at", "requirements"],
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Rustic AI API",
    "version": "0.0.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/catalog/blueprints/{blueprint_id}/guilds/preflight": {
      "post": {
        "operationId": "preflightGuildFromBlueprint",
        "parameters": [
          {
            "in": "path",
            "name": "blueprint_id",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/LaunchPreflightRequest"}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/LaunchPreflightResponse"}
              }
            },
            "description": "Launch requirements evaluated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HTTPValidationError"}
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preflight Guild From Blueprint",
        "tags": ["blueprints", "guilds"]
      }
    },
...
    "/oauth/callback": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "OAuth Callback",
        "operationId": "oauthCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An HTML page indicating success or failure.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}